+++ R/methods-epi_archive.R
@@ -133,16 +133,26 @@ epix_as_of <- function(x, version, min_time_value = -Inf, all_versions = FALSE,
   return(as_of_epi_df)
 }
 
-#' Get the latest snapshot from an `epi_archive` object.
+#' Get the latest snapshot from an `epi_archive` object
 #'
-#' The latest snapshot is the snapshot of the last known version.
+#' The latest snapshot is an `epi_df` snapshot of the data, as of the
+#' last recorded version in `x`.
 #'
 #' @param x An `epi_archive` object
-#' @return The latest snapshot from an `epi_archive` object
+#' @return The latest snapshot, in `epi_df` format
 #' @export
-epix_as_of_current <- function(x) {
+epix_as_of_latest <- function(x) {
   assert_class(x, "epi_archive")
   x %>% epix_as_of(.$versions_end)
+}
+
+#' @export
+epix_as_of_current <- function(x) {
+  lifecycle::deprecate_warn(
+    "0.13.0", "epix_as_of_current()", "epix_as_of_latest()",
+    details = 'Future versions of `epix_as_of_current` may remove this deprecation, but change its implementation to require "current" to be sometime recent, i.e., require there to be a nontrivial recent update to the data.'
+  )
+  epix_as_of_latest(x)
 }
 
 #' Set the `versions_end` attribute of an `epi_archive` object